feat: initial extraction of mouse.rs
